Ensure the save progress dialog is fully shown before starting request.
Due to the request often executing almost instantly, make sure the progress
dialog is fully shown before actually executing.  If it remains fast, this
should have little change except that the dialog will no longer flash.  If
it takes longer, this won't effect it much.

Also seems to kill the random failures with hiding said dialog.  If the
hiding failed, the code got stuck doing nothing, which is useless.