Make sure to return the promise when in a resolve function.
ngResource's returned from resolve will render immediately.  This is of course
broken, as the result may come later.  If the html page takes a similar amount
of time to be fetched as the result, it appears to complete fine.  However,
if the html page returns sooner, then the data won't be present causing
failures later on.