Add the ability to fetch all records in a given bucket to boltorm.
Boltorm now fetches all records.  Also adds tests and a memorydb implmentation,
as usual.

To handle slices, it unfortunately has to use the reflect package.  It is
all relatively straight forward.  All the tricky reflection code is moved
to a general handler to make life simpler.

The memory db also has to sort the outgoing records, as the map of records
is unsorted, and occasionally (though rarely, interestingly enough) causes
test failures.  Since the memorydb is meant for tests anyways, this isn't
a huge issue.