Fix up local javascript to match jshint settings.
Since grunt makes it easy to verify this, ensure jshint
settings and the code match.  jshintrc files are ajdusted to
my preferences, and the code updated to match.

Significantly, use double quotes for all strings (to match go
on the backend and json for communication), and don't use
semicolons except where needed, as Javascript's rules
surrouding them do make sense for the most part.