Have the session saving code be run no matter what happens.
It is not guaranteed that my http.ResponseWriter will have its WriteHeader
function be called if the code being serviced doesn't explicitly call it.
Thus ensure if some data is written that the session will still be saved.
Also, if nothing is written ensure the session saving is at least attempted.

This may mess up panics some, but oh well.